Make code compatible with pandas 1.3.x
Pandas 1.1.x would end up with one extra line of NaNs,
so a number of workarounds were used like `skipfooter=1`
to remove those lines.

These workarounds need to be removed, else we end up with missing data
or exceptions when trying to remove 'SKUPAJ' twice.

The `engine='openpyxl'` for pandas.read_excel is no longer needed
as this is now the default.